requirements.txt
streamlit 
deepface 
opencv-python 
mediapipe 
pillow